Launch now through jdee-maven plugin.
Previously, various dependencies (nrepl, clojure and so on) had to be
added to the project. These are now gone, so that it is possible to
launch a REPL on a "naked" project with no support for JDEE.

All the "jde" symbols have be changed to "jdee".

jde-interaction has changed name to jdee-live.